+++ homeassistant/components/owntracks/__init__.py
@@ -15,6 +15,7 @@ import homeassistant.helpers.config_validation as cv
 from homeassistant.setup import async_when_setup
 
 from .config_flow import CONF_SECRET
+from .messages import async_handle_message
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -50,7 +51,9 @@ CONFIG_SCHEMA = vol.Schema({
 async def async_setup(hass, config):
     """Initialize OwnTracks component."""
     hass.data[DOMAIN] = {
-        'config': config[DOMAIN]
+        'config': config[DOMAIN],
+        'devices': {},
+        'unsub': None,
     }
     if not hass.config_entries.async_entries(DOMAIN):
         hass.async_create_task(hass.config_entries.flow.async_init(
@@ -88,6 +91,10 @@ async def async_setup_entry(hass, entry):
     hass.async_create_task(hass.config_entries.async_forward_entry_setup(
         entry, 'device_tracker'))
 
+    hass.data[DOMAIN]['unsub'] = \
+        hass.helpers.dispatcher.async_dispatcher_connect(
+            DOMAIN, async_handle_message)
+
     return True
 
 
@@ -96,6 +103,8 @@ async def async_unload_entry(hass, entry):
     hass.components.webhook.async_unregister(entry.data[CONF_WEBHOOK_ID])
     await hass.config_entries.async_forward_entry_unload(
         entry, 'device_tracker')
+    hass.data[DOMAIN]['unsub']()
+
     return True
 
 
@@ -213,11 +222,13 @@ class OwnTracksContext:
 
         return True
 
-    async def async_see(self, **data):
+    @callback
+    def async_see(self, **data):
         """Send a see message to the device tracker."""
         raise NotImplementedError
 
-    async def async_see_beacons(self, hass, dev_id, kwargs_param):
+    @callback
+    def async_see_beacons(self, hass, dev_id, kwargs_param):
         """Set active beacons to the current location."""
         kwargs = kwargs_param.copy()
 
@@ -231,8 +242,13 @@ class OwnTracksContext:
             acc = device_tracker_state.attributes.get("gps_accuracy")
             lat = device_tracker_state.attributes.get("latitude")
             lon = device_tracker_state.attributes.get("longitude")
-            kwargs['gps_accuracy'] = acc
-            kwargs['gps'] = (lat, lon)
+
+            if lat is not None and lon is not None:
+                kwargs['gps'] = (lat, lon)
+                kwargs['gps_accuracy'] = acc
+            else:
+                kwargs['gps'] = None
+                kwargs['gps_accuracy'] = None
 
         # the battery state applies to the tracking device, not the beacon
         # kwargs location is the beacon's configured lat/lon
@@ -240,4 +256,4 @@ class OwnTracksContext:
         for beacon in self.mobile_beacons_active[dev_id]:
             kwargs['dev_id'] = "{}_{}".format(BEACON_DEV_ID, beacon)
             kwargs['host_name'] = beacon
-            await self.async_see(**kwargs)
+            self.async_see(**kwargs)
